// frontend/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Language = {
  ja: 'ja',
  en: 'en'
} as const

export type Language = (typeof Language)[keyof typeof Language]


export const Permission = {
  OWNER: 'OWNER',
  EDITOR: 'EDITOR',
  VIEWER: 'VIEWER'
} as const

export type Permission = (typeof Permission)[keyof typeof Permission]


export const NotificationType = {
  DIAGRAM_SHARED: 'DIAGRAM_SHARED',
  FOLDER_SHARED: 'FOLDER_SHARED',
  TEAM_SHARED: 'TEAM_SHARED',
  USER_INVITED: 'USER_INVITED'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]


export const Plan = {
  Free: 'Free',
  Plus: 'Plus',
  Pro: 'Pro'
} as const

export type Plan = (typeof Plan)[keyof typeof Plan]